Extract section capitalization scanner
