+++ src/scan/mod.rs
@@ -4,3 +4,4 @@
 pub(crate) mod duplicate_sections;
 pub(crate) mod section_capitalization;
 pub(crate) mod section_level;
+pub(crate) mod unordered_sections;
